+++ src/js/view/room/mouse/SelectAndDrag.mjs
@@ -87,8 +87,8 @@ export class SelectAndDrag extends MouseButtonHandler {
       this.dragging.piece = findPiece(target.id)
 
       this.dragging.style.zIndex = 999999999 // drag visually on top of everything
-      this.dragging.classList.add('dragging')
-      this.dragging.classList.add('dragging-hidden') // hide new item till it gets moved (1)
+      this.dragging.classList.add('is-dragging')
+      this.dragging.classList.add('is-dragging-hidden') // hide new item till it gets moved (1)
       target.parentNode.appendChild(this.dragging)
 
       this.dragging.startX = mousedown.clientX // no need to compensate, as we
@@ -100,7 +100,7 @@ export class SelectAndDrag extends MouseButtonHandler {
 
   drag (mousemove) {
     if (this.isDragging()) {
-      this.dragging.classList.remove('dragging-hidden') // we are moving now (1)
+      this.dragging.classList.remove('is-dragging-hidden') // we are moving now (1)
       moveNodeToSnapped(
         this.dragging,
         this.dragging.piece.x + mousemove.clientX - this.dragging.startX,